requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
attrs==24.2.0
certifi==2024.12.14
charset-normalizer==3.4.0
colorama==0.4.6
dnspython==2.7.0
frozenlist==1.5.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
MarkupSafe==3.0.2
motor==3.6.0
multidict==6.1.0
packaging==24.2
pluggy==1.5.0
propcache==0.2.1
pymongo==4.9.2
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
urllib3==2.2.3
websockets==14.1
yarl==1.18.3
